Move FakeFunction into test package
