+++ hack/generator/pkg/test/fake_function_test.go
@@ -3,7 +3,7 @@
  * Licensed under the MIT license.
  */
 
-package conversions
+package test
 
 import (
 	"github.com/dave/dst"
